requirements.txt
PyQt5==5.15.11
PyYAML==6.0
opencv-python
torch
ultralytics
